Alinear con codigo ya existente
