+++ package.json
@@ -21,5 +21,4 @@
     "collectCoverage": true,
     "coverageReporters": ["json", "html"]
     }
-  
 }
