+++ src/transaccion/TransaccionSimple.js
@@ -1,9 +1,17 @@
-const Transaction = require('./Transaccion');
+const TransaccionLeafAbstracta = require('./transaccionLeafAbstracta');
 
-class TransaccionSimple extends Transaction{
-    constructor(id, tkn, out, hash) {
-        super(id, tkn, out, hash);
-    }
+class TransaccionSimple extends TransaccionLeafAbstracta {
+  constructor(IN, OUT) {
+    super(IN, OUT);
+  }
+
+  agregarTransaccion(transaccion) {
+    throw new Error('No se pueden agregar transacciones a una transacción de tipo TransaccionSimple.');
+  }
+
+  obtenerNivel() {
+    return 0;
+  }
 }
 
 module.exports = TransaccionSimple;
