+++ src/transaccion/transaccionLeafAbstracta.js
@@ -1,0 +1,9 @@
+const TransaccionAbstracta = require('./transaccionAbstracta');
+
+class TransaccionLeafAbstracta extends TransaccionAbstracta {
+  constructor(IN, OUT) {
+    super();
+  }
+}
+
+module.exports = TransaccionLeafAbstracta;
