--- src/transaccionLeafAbstracta.js
@@ -1,9 +1,0 @@
-const TransaccionAbstracta = require('./transaccionAbstracta');
-
-class TransaccionLeafAbstracta extends TransaccionAbstracta {
-  constructor(IN, OUT) {
-    super();
-  }
-}
-
-module.exports = TransaccionLeafAbstracta;
